refactor: separator ProfileDialog
